Replace int with isize

+++ src/sdl2_ttf/lib.rs
@@ -62,20 +62,20 @@ bitflags!(flags FontStyle : c_int {
 
 #[deriving(Show, PartialEq, FromPrimitive)]
 pub enum Hinting {
-    HintingNormal = ffi::TTF_HINTING_NORMAL as int,
-    HintingLight  = ffi::TTF_HINTING_LIGHT  as int,
-    HintingMono   = ffi::TTF_HINTING_MONO   as int,
-    HintingNone   = ffi::TTF_HINTING_NONE   as int
+    HintingNormal = ffi::TTF_HINTING_NORMAL as isize,
+    HintingLight  = ffi::TTF_HINTING_LIGHT  as isize,
+    HintingMono   = ffi::TTF_HINTING_MONO   as isize,
+    HintingNone   = ffi::TTF_HINTING_NONE   as isize
 }
 
 /// Glyph Metrics
 #[deriving(PartialEq, Clone, Show)]
 pub struct GlyphMetrics {
-    pub minx: int,
-    pub maxx: int,
-    pub miny: int,
-    pub maxy: int,
-    pub advance: int
+    pub minx: isize,
+    pub maxx: isize,
+    pub miny: isize,
+    pub maxy: isize,
+    pub advance: isize
 }
 
 /// Returns the version of the dynamically linked SDL_ttf library
@@ -134,7 +134,7 @@ impl Font {
         Font { raw: raw, owned: owned }
     }
 
-    pub fn from_file(filename: &Path, ptsize: int) -> SdlResult<Font> {
+    pub fn from_file(filename: &Path, ptsize: isize) -> SdlResult<Font> {
         //! Load file for use as a font, at ptsize size.
         unsafe {
             let raw = ffi::TTF_OpenFont(filename.to_c_str().unwrap(), ptsize as c_int);
@@ -146,7 +146,7 @@ impl Font {
         }
     }
 
-    pub fn from_file_index(filename: &Path, ptsize: int, index: int) -> SdlResult<Font> {
+    pub fn from_file_index(filename: &Path, ptsize: isize, index: isize) -> SdlResult<Font> {
         //! Load file, face index, for use as a font, at ptsize size.
         unsafe {
             let raw = ffi::TTF_OpenFontIndex(filename.to_c_str().unwrap(), ptsize as c_int, index as c_long);
@@ -173,14 +173,14 @@ impl Font {
         }
     }
 
-    pub fn get_outline(&self) -> int {
+    pub fn get_outline(&self) -> isize {
         //! Get font outline width.
         unsafe {
-            ffi::TTF_GetFontOutline(self.raw) as int
+            ffi::TTF_GetFontOutline(self.raw) as isize
         }
     }
 
-    pub fn set_outline(&mut self, outline: int) {
+    pub fn set_outline(&mut self, outline: isize) {
         //! Set font outline width.
         unsafe {
             ffi::TTF_SetFontOutline(self.raw, outline as c_int)
@@ -215,38 +215,38 @@ impl Font {
         }
     }
 
-    pub fn height(&self) -> int {
+    pub fn height(&self) -> isize {
         //! Get font maximum total height.
         unsafe {
-            ffi::TTF_FontHeight(self.raw) as int
+            ffi::TTF_FontHeight(self.raw) as isize
         }
     }
 
-    pub fn ascent(&self) -> int {
+    pub fn ascent(&self) -> isize {
         //! Get font highest ascent (height above base).
         unsafe {
-            ffi::TTF_FontAscent(self.raw) as int
+            ffi::TTF_FontAscent(self.raw) as isize
         }
     }
 
-    pub fn descent(&self) -> int {
+    pub fn descent(&self) -> isize {
         //! Get font lowest descent (height below base).
         unsafe {
-            ffi::TTF_FontDescent(self.raw) as int
+            ffi::TTF_FontDescent(self.raw) as isize
         }
     }
 
-    pub fn line_skip(&self) -> int {
+    pub fn line_skip(&self) -> isize {
         //! Get font recommended line spacing.
         unsafe {
-            ffi::TTF_FontLineSkip(self.raw) as int
+            ffi::TTF_FontLineSkip(self.raw) as isize
         }
     }
 
-    pub fn faces(&self) -> int {
+    pub fn faces(&self) -> isize {
         //! Get the number of faces in a font.
         unsafe {
-            ffi::TTF_FontFaces(self.raw) as int
+            ffi::TTF_FontFaces(self.raw) as isize
         }
     }
 
@@ -282,14 +282,14 @@ impl Font {
         }
     }
 
-    pub fn index_of_char(&self, ch: char) -> Option<uint> {
+    pub fn index_of_char(&self, ch: char) -> Option<usize> {
         //! Get individual font glyph availability.
         unsafe {
             let ret = ffi::TTF_GlyphIsProvided(self.raw, ch as u16);
             if ret == 0 {
                 None
             } else {
-                Some(ret as uint)
+                Some(ret as usize)
             }
         }
     }
@@ -308,13 +308,13 @@ impl Font {
         if ret != 0 {
             None
         } else {
-            Some(GlyphMetrics { minx: minx as int, maxx: maxx as int,
-                                miny: miny as int, maxy: maxy as int,
-                                advance: advance as int })
+            Some(GlyphMetrics { minx: minx as isize, maxx: maxx as isize,
+                                miny: miny as isize, maxy: maxy as isize,
+                                advance: advance as isize })
         }
     }
 
-    pub fn size_of_bytes(&self, text: &[u8]) -> SdlResult<(int, int)> {
+    pub fn size_of_bytes(&self, text: &[u8]) -> SdlResult<(isize, isize)> {
         //! Get size of LATIN1 text string as would be rendered.
         let w = 0;
         let h = 0;
@@ -326,11 +326,11 @@ impl Font {
         if ret != 0 {
             Err(get_error())
         } else {
-            Ok((w as int, h as int))
+            Ok((w as isize, h as isize))
         }
     }
 
-    pub fn size_of_str(&self, text: &str) -> SdlResult<(int, int)> {
+    pub fn size_of_str(&self, text: &str) -> SdlResult<(isize, isize)> {
         //! Get size of UTF8 text string as would be rendered.
         let w = 0;
         let h = 0;
@@ -342,7 +342,7 @@ impl Font {
         if ret != 0 {
             Err(get_error())
         } else {
-            Ok((w as int, h as int))
+            Ok((w as isize, h as isize))
         }
     }
 
@@ -471,13 +471,13 @@ impl Font {
 /// Loader trait for RWops
 pub trait LoaderRWops {
     /// Load src for use as a font.
-    fn load_font(&self, ptsize: int) -> SdlResult<Font>;
+    fn load_font(&self, ptsize: isize) -> SdlResult<Font>;
     /// Load src for use as a font.
-    fn load_font_index(&self, ptsize: int, index: int) -> SdlResult<Font>;
+    fn load_font_index(&self, ptsize: isize, index: isize) -> SdlResult<Font>;
 }
 
 impl LoaderRWops for RWops {
-    fn load_font(&self, ptsize: int) -> SdlResult<Font> {
+    fn load_font(&self, ptsize: isize) -> SdlResult<Font> {
         let raw = unsafe {
             ffi::TTF_OpenFontRW(self.raw(), 0, ptsize as c_int)
         };
@@ -487,7 +487,7 @@ impl LoaderRWops for RWops {
             Ok(Font::from_ll(raw, true))
         }
     }
-    fn load_font_index(&self, ptsize: int, index: int) -> SdlResult<Font> {
+    fn load_font_index(&self, ptsize: isize, index: isize) -> SdlResult<Font> {
         let raw = unsafe {
             ffi::TTF_OpenFontIndexRW(self.raw(), 0, ptsize as c_int, index as c_long)
         };
